third_party/WebKit/LayoutTests/fast/dom/documentmarker-add-adjacent-text-expected.txt

Checks that two adjacent TextMatch markers are not being merged.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.markerCountForNode(elt, "textmatch") is 2
PASS successfullyParsed is true

TEST COMPLETE

text
